Move coverFetchers to cover_fetchers
