+++ etl-pipeline/managers/cover_fetchers/__init__.py
