+++ etl-pipeline/managers/cover_fetchers/abstractFetcher.py
