+++ etl-pipeline/managers/cover_fetchers/contentCafeFetcher.py
@@ -4,7 +4,7 @@ import requests
 from requests.exceptions import ReadTimeout
 
 from logger import create_log
-from managers.coverFetchers.abstractFetcher import AbstractFetcher
+from managers.cover_fetchers.abstractFetcher import AbstractFetcher
 
 logger = create_log(__name__)
