+++ etl-pipeline/managers/cover_fetchers/googleBooksFetcher.py
@@ -3,7 +3,7 @@ import requests
 from requests.exceptions import ReadTimeout, HTTPError
 
 from logger import create_log
-from managers.coverFetchers.abstractFetcher import AbstractFetcher
+from managers.cover_fetchers.abstractFetcher import AbstractFetcher
 
 logger = create_log(__name__)
