+++ etl-pipeline/managers/cover_fetchers/hathiFetcher.py
@@ -3,7 +3,7 @@ import requests
 from requests.exceptions import ReadTimeout, HTTPError
 from requests_oauthlib import OAuth1
 
-from managers.coverFetchers.abstractFetcher import AbstractFetcher
+from managers.cover_fetchers.abstractFetcher import AbstractFetcher
 from logger import create_log
 
 logger = create_log(__name__)
