+++ etl-pipeline/tests/unit/test_fetchers_abstract.py
@@ -1,6 +1,6 @@
 import pytest
 
-from managers.coverFetchers.abstractFetcher import AbstractFetcher
+from managers.cover_fetchers.abstractFetcher import AbstractFetcher
 
 
 class TestAbstractFetcher:
